+++ src/test/org/x96/sys/foundation/cs/ast/book/passage/pattern/core/SegmentTest.java
@@ -13,7 +13,7 @@ class SegmentTest {
     @Test
     void happy(){
         Segment s = new Segment(new Axis[]{
-                new RangeHex(new Hexadecimal(0x1), new Hexadecimal(0x2)),
+                new RangeHex(new Hexadecimal((byte) 0x1), new Hexadecimal((byte) 0x2)),
                 new Word("cs".getBytes()),
         }, Optional.empty());
         assertEquals("segment", s.primor());
@@ -29,10 +29,10 @@ class SegmentTest {
     @Test
     void happyRecursive(){
         Segment s = new Segment(new Axis[]{
-                new RangeHex(new Hexadecimal(0x1), new Hexadecimal(0x2)),
+                new RangeHex(new Hexadecimal((byte) 0x1), new Hexadecimal((byte) 0x2)),
                 new Word("cs".getBytes()),
                 new Segment(new Axis[]{
-                        new RangeHex(new Hexadecimal(0x11), new Hexadecimal(0x22)),
+                        new RangeHex(new Hexadecimal((byte) 0x11), new Hexadecimal((byte) 0x22)),
                         new Word("sc".getBytes())
                 }, Optional.empty())
         }, Optional.empty());
@@ -54,10 +54,10 @@ class SegmentTest {
     @Test
     void happyRecursiveQuantifier(){
         Segment s = new Segment(new Axis[]{
-                new RangeHex(new Hexadecimal(0x1), new Hexadecimal(0x2)),
+                new RangeHex(new Hexadecimal((byte) 0x1), new Hexadecimal((byte) 0x2)),
                 new Word("cs".getBytes()),
                 new Segment(new Axis[]{
-                        new RangeHex(new Hexadecimal(0x11), new Hexadecimal(0x22)),
+                        new RangeHex(new Hexadecimal((byte) 0x11), new Hexadecimal((byte) 0x22)),
                         new Word("sc".getBytes())
                 }, Optional.empty())
         }, Optional.of(new ZeroOrMore("*".getBytes()[0])));
